# src/server/api/swagger.json
{
  "swagger": "2.0",
  "produces": [
    "application/json"
  ],
  "info": {
    "title": "conversationdiscovery",
    "description": "An API",
    "version": "1.0"
  },
  "basePath": "/conversationdiscovery",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "in": "path",
      "description": "A client id, e.g. \"247inc\".",
      "required": true,
      "type": "string"
    },
    "applicationId": {
      "name": "applicationId",
      "in": "path",
      "description": "An app id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "apiRequestBodyNode": {
      "name": "body",
      "description": "body for nodes POST and PATCH operations",
      "in": "body",
      "required": true,
      "schema": {
        "$ref": "#/definitions/ApiRequestBodyNode"
      }
    },
    "accountId": {
      "name": "accountId",
      "in": "path",
      "description": "An account id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "runId": {
      "name": "runId",
      "in": "path",
      "description": "Run id of a run",
      "required": true,
      "type": "string",
      "format": "uuid"
    },
    "clusterId": {
      "name": "clusterId",
      "in": "path",
      "description": "Cluster identifier",
      "required": true,
      "type": "string"
    },
...
    "projectId": {
      "name": "projectId",
      "in": "path",
      "description": "id of the project",
      "required": true,
      "type": "string"
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "The request was invalid. For example, the value of the initial parameter was invalid",
      "schema": {
        "type": "string"
      }
    },
    "NotFound": {
      "description": "The URI requested is invalid or the resource requested, such as a user, does not exist. Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
...
    "InternalServerError": {
      "description": "Internal server error",
      "schema": {
        "type": "string"
      }
    },
...
  },
  "paths": {
...
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects/{projectId}/runs/{runId}/clusters/{clusterId}/intents": {
      "patch": {
        "summary": "update given node",
        "description": "400: MUST NOT create a cycle; response MUST match message type; MUST NOT use nodeId body parameter",
        "x-swagger-router-controller": "nodes",
        "operationId": "updateIntent",
        "tags": [
          "node"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/clusterId"
          },
          {
            "$ref": "#/parameters/apiRequestBodyNode"
          }
        ],
        "responses": {
          "201": {
            "description": "new child node, or updated child node with additional parent id",
            "schema": {
              "$ref": "#/definitions/Node"
            },
            "headers": {
              "ETag": {
                "description": "identifier for a specific version of a node resource",
                "type": "string"
              }
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "summary": "create node, or connect existing node, as child of given node in path",
        "description": "400: MUST NOT create a cycle; response MUST match message type; connecting uuid MUST exist; MUST NOT combine nodeId with non-nodeId body parameters",
        "x-swagger-router-controller": "nodes",
        "operationId": "createIntent",
        "tags": [
          "node"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/clusterId"
          },
          {
            "$ref": "#/parameters/apiRequestBodyNode"
          }
        ],
        "responses": {
          "200": {
            "description": "updated node",
            "schema": {
              "$ref": "#/definitions/Node"
            },
            "headers": {
              "ETag": {
                "description": "identifier for a specific version of a node resource",
                "type": "string"
              }
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "get": {
        "summary": "Get all the intents for a given app and client",
        "x-swagger-router-controller": "nodes",
        "operationId": "getIntents",
        "tags": [
          "node"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "$ref": "#/parameters/clusterId"
          }
        ],
        "responses": {
          "200": {
            "description": "list of nodes",
            "schema": {
              "$ref": "#/definitions/ApiResponsePagedGetNodes"
            },
            "headers": {
              "ETag": {
                "description": "identifier for a specific version of a node resource",
                "type": "string"
              }
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApiRequestBodyNode": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "nodeId": {
          "description": "id of existing node; USE ALONE, not in combination with other properties",
          "type": "string",
          "example": "12345678-abcd-1234-8765-fedcba987654"
        },
        "nodeName": {
          "description": "user-defined node name; REQUIRED if not using nodeId for existing node",
          "type": "string",
          "pattern": "^[a-zA-Z0-9_]+$",
          "example": "Cancel_reservation"
        },
        "intentModelMap": {
          "$ref": "#/definitions/IntentModelMap"
        },
        "responses": {
          "$ref": "#/definitions/Responses"
        },
        "utterances": {
          "$ref": "#/definitions/Utterances"
        }
      }
    },
    "ApiResponsePagedGetNodes": {
      "description": "add pagination to getNodes query",
      "type": "object",
      "properties": {
        "paging": {
          "$ref": "#/definitions/Pagination"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Node"
          }
        }
      }
    },
    "Node": {
      "type": "object",
      "required": [
        "nodeName"
      ],
      "properties": {
        "nodeName": {
          "description": "user-defined node name",
          "type": "string",
          "example": "Cancel_reservation"
        },
        "nodeId": {
          "description": "system-generated node identifier",
          "type": "string",
          "readOnly": true,
          "example": "420f9b81-450c-46d7-85d0-a23f5c4d9445"
        },
        "normalizedName": {
          "description": "system-derived unique node name suitable for NL modeling",
          "type": "string",
          "readOnly": true,
          "example": "Reservation_Cancel"
        },
        "operation": {
          "description": "system-maintained record of change to this node since most recent build",
          "type": "string",
          "readOnly": true,
          "enum": [
            "noop",
            "create",
            "update",
            "delete"
          ]
        },
        "context": {
          "$ref": "#/definitions/Context"
        },
        "intentModelMap": {
          "$ref": "#/definitions/IntentModelMap"
        },
        "responses": {
          "$ref": "#/definitions/Responses"
        },
        "utterances": {
          "$ref": "#/definitions/Utterances"
        },
        "metadata": {
          "$ref": "#/definitions/OptionsMetadata"
        },
        "defects": {
          "$ref": "#/definitions/Defects"
        }
      }
    },
    "Context": {
      "description": "system-maintained connection between nodes",
      "type": "object",
      "readOnly": true,
      "properties": {
        "inputContext": {
          "description": "sorted list of parent nodes",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "outputContext": {
          "description": "sorted list of child nodes",
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "example": {
        "inputContext": [
          "420f9b81-450c-46d7-85d0-a23f5c4d944"
        ],
        "outputContext": [
          "98ddaf3f-ce91-4d07-b1d0-cee12c7a9f55"
        ]
      }
    },
    "IntentModelMap": {
      "description": "map from intent values to destination nodes",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "example": {
        "app.match": "420f9b81-450c-46d7-85d0-a23f5c4d944",
        "no.give": "98ddaf3f-ce91-4d07-b1d0-cee12c7a9f55"
      }
    },
    "Responses": {
      "description": "ordered list(s) of responses for all MIME types and locales",
      "type": "array",
      "minItems": 1,
      "items": {
        "$ref": "#/definitions/ResponseItem"
      }
    },
    "ResponseItem": {
      "description": "base item in responses array",
      "type": "object",
      "discriminator": "messageType",
      "required": [
        "event",
        "messageType"
      ],
      "properties": {
        "event": {
          "description": "event when this response is relevant",
          "type": "string",
          "enum": [
            "success",
            "retry"
          ],
          "example": "success"
        },
        "messageType": {
          "description": "rendering and handling pattern",
          "type": "string",
          "enum": [
            "api",
            "escalate",
            "in.choose",
            "in.form",
            "in.nlu",
            "out.form",
            "out.msg",
            "script",
            "terminate"
          ],
          "example": "out.msg"
        },
        "mimeType": {
          "description": "MIME type of response (default is text/html)",
          "type": "string",
          "enum": [
            "text/html"
          ],
          "example": "text/html"
        },
        "locale": {
          "description": "IETF language tag with hyphen (default is en-US)",
          "type": "string",
          "enum": [
            "en-US"
          ],
          "example": "en-US"
        },
        "offset": {
          "description": "zero-based index for sorting all items with same mimeType-locale",
          "type": "integer",
          "example": 0
        }
      }
    },
...
    "Utterances": {
      "description": "weighted NL text",
      "type": "array",
      "minItems": 0,
      "items": {
        "type": "object",
        "properties": {
          "utterance": {
            "description": "NL text",
            "type": "string",
            "example": "i want to cancel a reservation"
          },
          "count": {
            "description": "weight of utterance in NL model",
            "type": "integer",
            "example": 2
          },
          "utteranceId": {
            "description": "utteranceId",
            "type": "string",
            "example": "utteranceId1"
          }
        }
      }
    },
    "Defects": {
      "description": "defects/warnings associated with the node",
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "defectType": {
            "description": "code for defect type, required for all defects",
            "type": "string",
            "enum": [
              "MANDATORY_FIELDS_MISSING",
              "DUPLICATE_UTTERANCE",
              "RESPONSE_OFFSET_ERROR"
            ]
          },
          "fields": {
            "description": "field list of missing mandatory fields, used with MANDATORY_FIELDS_MISSING",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "utterances": {
            "description": "utterances which are duplicated, used with DUPLICATE_UTTERANCE",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "responses": {
            "description": "stringified responses with mimeType and offset values which are erroneous, used with RESPONSE_OFFSET_ERROR",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "OptionsMetadata": {
      "description": "Usage constraints and guidelines on given node",
      "type": "object",
      "properties": {
        "allow": {
          "description": "http operations allowed on this node",
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "GET",
              "HEAD",
              "OPTIONS",
              "PATCH",
              "POST"
            ]
          },
          "default": [
            "OPTIONS",
            "GET",
            "PATCH"
          ]
        },
        "ETag": {
          "description": "identifier for a specific version of a node resource",
          "type": "string"
        },
        "x-tfs-readonly": {
          "description": "properties defined as readOnly=false but cannot be modified",
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "nodeName",
              "intentModelMap",
              "responses",
              "utterances"
            ]
          },
          "default": []
        }
      }
    },
    "Pagination": {
      "description": "offset-based pagination of large collection",
      "type": "object",
      "properties": {
        "offset": {
          "description": "zero-based index into array of items (default is 0)",
          "type": "integer",
          "example": 0
        },
        "limit": {
          "description": "maximum number of items in page response (default is 25)",
          "type": "integer",
          "example": 25
        },
        "total": {
          "description": "size of collection",
          "type": "integer",
          "example": 59
        }
      }
    }
  }
}